+++ src/lib-public/generators/FA3/RachunekBankowy.ts
@@ -8,8 +8,9 @@ import {
   makeBreakable,
 } from '../../../shared/PDF-functions';
 import FormatTyp from '../../../shared/enums/common.enum';
+import { TypRachunkowWlasnych } from '../../../shared/consts/FA.const';
 import { RachunekBankowy } from '../../types/fa3.types';
-import { getTypRachunkowWlasnych } from '../../../shared/generators/common/functions';
+import { translateMap } from '../../../shared/generators/common/functions';
 
 export const generujRachunekBankowy = (accounts?: RachunekBankowy[], title?: string): Content[] => {
   const result: Content[] = [];
@@ -35,7 +36,7 @@ export const generujRachunekBankowy = (accounts?: RachunekBankowy[], title?: str
     ]);
     table.push([
       formatText('Rachunek własny banku', FormatTyp.GrayBoldTitle),
-      formatText(makeBreakable(getTypRachunkowWlasnych(account.RachunekWlasnyBanku), 20), FormatTyp.Default),
+      formatText(makeBreakable(translateMap(account.RachunekWlasnyBanku, TypRachunkowWlasnych), 20), FormatTyp.Default),
     ]);
     table.push([
       formatText('Nazwa banku', FormatTyp.GrayBoldTitle),