+++ src/lib-public/generators/FA3/Zamowienie.ts
@@ -9,7 +9,7 @@ import {
   getValue,
 } from '../../../shared/PDF-functions';
 import { HeaderDefine } from '../../../shared/types/pdf-types';
-import { TRodzajFaktury } from '../../../shared/consts/const';
+import { TRodzajFaktury } from '../../../shared/consts/FA.const';
 import { Zamowienie } from '../../types/fa3.types';
 import FormatTyp, { Position } from '../../../shared/enums/common.enum';
 import { ZamowienieKorekta } from '../../enums/invoice.enums';