+++ src/shared/PDF-functions.ts
@@ -10,9 +10,16 @@ import {
   TableCell,
   TDocumentDefinitions,
 } from 'pdfmake/interfaces';
-import { DEFAULT_TABLE_LAYOUT, Kraj, TStawkaPodatku_FA1, TStawkaPodatku_FA2, TStawkaPodatku_FA3 } from './consts/const';
+import {
+  DEFAULT_TABLE_LAYOUT,
+  FormaPlatnosci,
+  TStawkaPodatku_FA1,
+  TStawkaPodatku_FA2,
+  TStawkaPodatku_FA3,
+} from './consts/FA.const';
+import { Kraj } from './consts/const';
 import { TStawkaPodatku_FARR } from './consts/FARR.const';
-import { formatDateTime, formatTime, getFormaPlatnosciString } from './generators/common/functions';
+import { formatDateTime, formatTime, translateMap } from './generators/common/functions';
 import { HeaderDefine, PdfFP, PdfOptionField } from './types/pdf-types';
 import { FP } from '../lib-public/types/fa3.types';
 import { DifferentValues, FilteredKeysOfValues, TypesOfValues } from './types/universal.types';
@@ -130,7 +137,7 @@ function formatValue(
       result.text = formatTime(value as string);
       break;
     case FormatTyp.FormOfPayment:
-      result.text = getFormaPlatnosciString({ _text: value as string });
+      result.text = translateMap(value as string, FormaPlatnosci);
       break;
     case FormatTyp.Boolean:
       result.text = (value as string) === '1' ? Answer.TRUE : Answer.FALSE;