+++ src/shared/generators/common/functions.ts
@@ -1,15 +1,3 @@
-import {
-  FA1RolaPodmiotu3,
-  FA2RolaPodmiotu3,
-  FA3RolaPodmiotu3,
-  FormaPlatnosci,
-  RodzajTransportu,
-  TRolaPodmiotuUpowaznionegoFA1,
-  TRolaPodmiotuUpowaznionegoFA2,
-  TRolaPodmiotuUpowaznionegoFA3,
-  TypLadunku,
-  TypRachunkowWlasnych,
-} from '../../consts/const';
 import { FP as FP2 } from '../../../lib-public/types/fa2.types';
 
 export function translateMap(value: FP2 | string | undefined, map: Record<string, string>): string {
@@ -20,62 +8,6 @@ export function translateMap(value: FP2 | string | undefined, map: Record<string
   }
 
   return map[valueToTranslate] ?? '';
-}
-
-export function getRolaString(rola: FP2 | undefined, FA: 1 | 2 | 3): string {
-  if (!rola || !rola._text) {
-    return '';
-  }
-  switch (FA) {
-    case 1:
-      return FA1RolaPodmiotu3[rola._text as keyof typeof FA1RolaPodmiotu3] ?? '';
-    case 2:
-      return FA2RolaPodmiotu3[rola._text as keyof typeof FA2RolaPodmiotu3] ?? '';
-    case 3:
-      return FA3RolaPodmiotu3[rola._text as keyof typeof FA3RolaPodmiotu3] ?? '';
-  }
-}
-
-export function getRolaUpowaznionegoString(rola: FP2 | undefined, FA: 1 | 2 | 3): string {
-  if (!rola || !rola._text) {
-    return '';
-  }
-  switch (FA) {
-    case 1:
-      return TRolaPodmiotuUpowaznionegoFA1[rola._text].split('-')[0] ?? '';
-    case 2:
-      return TRolaPodmiotuUpowaznionegoFA2[rola._text].split('-')[0] ?? '';
-    case 3:
-      return TRolaPodmiotuUpowaznionegoFA3[rola._text].split('-')[0] ?? '';
-  }
-}
-
-export function getFormaPlatnosciString(formaPlatnosci: FP2 | undefined): string {
-  if (!formaPlatnosci || !formaPlatnosci._text) {
-    return '';
-  }
-  return FormaPlatnosci[formaPlatnosci._text as keyof typeof FormaPlatnosci] ?? '';
-}
-
-export function getRodzajTransportuString(rodzajTransportu: FP2 | undefined): string {
-  if (!rodzajTransportu || !rodzajTransportu._text) {
-    return '';
-  }
-  return RodzajTransportu[rodzajTransportu._text as keyof typeof RodzajTransportu] ?? '';
-}
-
-export function getOpisTransportuString(opisTransportu: FP2 | undefined): string {
-  if (!opisTransportu || !opisTransportu._text) {
-    return '';
-  }
-  return TypLadunku[opisTransportu._text as keyof typeof TypLadunku] ?? '';
-}
-
-export function getTypRachunkowWlasnych(typRachonkowWlasnych: FP2 | undefined): string {
-  if (!typRachonkowWlasnych || !typRachonkowWlasnych._text) {
-    return '';
-  }
-  return TypRachunkowWlasnych[typRachonkowWlasnych._text as keyof typeof TypRachunkowWlasnych] ?? '';
 }
 
 export function formatDateTime(data?: string, withoutSeconds?: boolean, withoutTime?: boolean): string {